Moved mod processing outside the worker
